first two steps of new dataset modal
