+++ DashAI/front/src/components/datasets/NewDatasetModal.jsx
@@ -13,33 +13,44 @@ import {
   Typography,
   StepButton,
 } from "@mui/material";
-import SchemaList from "../SchemaList";
+import ItemsList from "./ItemsList";
 
 const steps = [
   { name: "selectTask", label: "select Task" },
   { name: "selectDataloader", label: "Select how to load" },
-  { name: "configureDataset", label: "Configure your dataset" },
-  { name: "uploadDataset", label: "Upload your dataset" },
+  { name: "uploadDataset", label: "Configure and upload your dataset" },
 ];
 
 const defaultNewDataset = {
   task_name: "",
+  dataloader_name: "",
 };
 
+/**
+ * This component renders a modal that takes the user through the process of uploading a new dataset.
+ * @param {bool} open true to open the modal, false to close it
+ * @param {function} setOpen function to modify the value of open
+ */
 function NewDatasetModal({ open, setOpen }) {
   const [activeStep, setActiveStep] = useState(0);
-  // const [nextEnabled, setNextEnabled] = useState(false);
+  const [nextEnabled, setNextEnabled] = useState(false);
   const [newDataset, setNewDataset] = useState(defaultNewDataset);
 
   const handleCloseDialog = () => {
     setActiveStep(0);
     setOpen(false);
+    setNewDataset(defaultNewDataset);
+    setNextEnabled(false);
+  };
+
+  const handleStepButton = (stepIndex) => () => {
+    setActiveStep(stepIndex);
   };
 
   const handleNextButton = () => {
     if (activeStep < steps.length) {
       setActiveStep(activeStep + 1);
-      // setNextEnabled(false);
+      setNextEnabled(false);
     } else {
       handleCloseDialog();
     }
@@ -64,13 +75,13 @@ function NewDatasetModal({ open, setOpen }) {
         sx: { minHeight: "80vh" },
       }}
     >
+      {/* Title */}
       <DialogTitle id="new-experiment-dialog-title">
         <Grid container direction={"row"} alignItems={"center"}>
           <Grid item xs={12} md={3}>
             <Typography
               variant="h6"
               component={"h3"}
-              // align={matches ? "center" : "left"}
               sx={{ mb: { sm: 2, md: 0 } }}
             >
               New dataset
@@ -88,7 +99,7 @@ function NewDatasetModal({ open, setOpen }) {
                   completed={activeStep > index}
                   disabled={activeStep < index}
                 >
-                  <StepButton color="inherit" onClick={() => {}}>
+                  <StepButton color="inherit" onClick={handleStepButton(index)}>
                     {step.label}
                   </StepButton>
                 </Step>
@@ -97,32 +108,43 @@ function NewDatasetModal({ open, setOpen }) {
           </Grid>
         </Grid>
       </DialogTitle>
+
+      {/* Main content - steps */}
       <DialogContent dividers>
         {activeStep === 0 && (
-          <SchemaList
-            schemaName="tasks"
+          <ItemsList
+            itemsType="tasks"
             itemsName="task"
             newDataset={newDataset}
             setNewDataset={setNewDataset}
+            setNextEnabled={setNextEnabled}
           />
         )}
         {activeStep === 1 && (
-          <SchemaList
-            schemaName="TabularClassificationTask"
-            itemsName="Way to upload your data"
+          <ItemsList
+            itemsType="dataloaders"
+            itemsName="way to upload your data"
             newDataset={newDataset}
             setNewDataset={setNewDataset}
+            setNextEnabled={setNextEnabled}
           />
         )}
+        {/* TODO: step 3 configuration and upload file */}
+        {activeStep === 2 && <div />}
       </DialogContent>
+
+      {/* Actions - Back and Next */}
       <DialogActions>
         <ButtonGroup size="large">
-          <Button onClick={handleBackButton}>Back</Button>
+          <Button onClick={handleBackButton}>
+            {activeStep === 0 ? "Close" : "Back"}
+          </Button>
           <Button
             onClick={handleNextButton}
             autoFocus
             variant="contained"
             color="primary"
+            disabled={!nextEnabled}
           >
             Next
           </Button>
